test: reorg the cupy tests for clarity
