+++ tests/test_cupy_gpu/test_fourier_cupy.py
